requirements.txt
asgiref==3.11.0
Django==6.0.1
numpy==2.4.1
opencv-python==4.13.0.90
pillow==12.1.0
pyzbar==0.1.9
qrcode==8.2
sqlparse==0.5.5
